Removed USE_AUTOEQ macro. In-game time related code and structures have been moved to weather.cpp from utils.
