+++ src/engine/ui/color.h
@@ -9,6 +9,8 @@
 #ifndef COLOR_H_
 #define COLOR_H_
 
+#include "engine/structs/structs.h"
+
 extern const char *kColorNrm;
 extern const char *kColorRed;
 extern const char *kColorGrn;
@@ -28,7 +30,9 @@ extern const char *kColorBoldMag;
 extern const char *kColorBoldCyn;
 extern const char *kColorBoldWht;
 
+std::size_t strlen_no_colors(const char *str);
 int proc_color(char *inbuf);
+char *colored_name(const char *str, std::size_t len, bool left_align = false);
 const char *GetWarmValueColor(int current, int max);
 const char *GetColdValueColor(int current, int max);
