+++ src/engine/ui/table_wrapper.h
@@ -3,20 +3,20 @@
  \date 18.02.2022.
  \brief Обертка-помощник модуля таблиц.
  \details Обертка модуля таблиц. Цель в том, чтобы собрать в одном месте процедуры оформления таблиц: не стоит
- в каждом случае формировать окраску, отступы и цвета вручную, достаточно определить несколько основных типов: магазины,
+ в каждом случае формировать окраску, отступы и цвета вручную. Достаточно определить несколько основных типов: магазины,
  "красивые" таблицы, вроде "счет все" и "моястатистика", просто статистика, типа списка клана, и  служебные, для вывода
  информации иммморталам. Вторая задача - избежать повторения в формировании каждой таблицы кода выбора режима
  оформления для слабовидящих игроков. Наконец, третья задача - облегчить смену модуля таблиц при возникновении такой
  необходимости.
 
  Следует иметь в виду, что стиль таблицы FT_EMPTY2_STYLE в используемой сейчас библиотеке libfort слегка изменен,
- чтобы соответствовать нуждам незрячих игроков. Окнкретно - добавлен разделитель столбцов и заголовков столбцов
+ чтобы соответствовать нуждам незрячих игроков. Конкретно - добавлен разделитель столбцов и заголовков столбцов
  в виде двоточия. При обновлении библиотеки стиль следует снова поменять (к сожалению, добавить соверщенно новый стиль
  в библиотеку несколько сложнее, а настройку оформления стиля она не поддерживает).
 
- Также нуэно учитывать, что в настоязий момент библиотека не подерживает и escape-последовательности цветов telnet.
- Чтобы цвета все-таки отображались, при9лось вручную заменть массив последовательностей. Что опять-таки приводит
- к необходимости повторить эту работу при обновлении библиотеки (при условии, что в новых версияъ не будет добавлена
+ Также нужно учитывать, что в настоящий момент библиотека не подерживает и escape-последовательности цветов telnet.
+ Чтобы цвета все-таки отображались, пришлось вручную заменть массив последовательностей. Что опять-таки приводит
+ к необходимости повторить эту работу при обновлении библиотеки (при условии, что в новых версиях не будет добавлена
  поддержка telnet, разумеется).
  */
