+++ src/gameplay/mechanics/sight.h
@@ -27,6 +27,7 @@ char *find_exdesc(const char *word, const ExtraDescription::shared_ptr &list);
 const char *show_obj_to_char(ObjData *object, CharData *ch, int mode, int show_state, int how);
 void obj_info(CharData *ch, ObjData *obj, char buf[kMaxStringLength]);
 void print_zone_info(CharData *ch);
+const char *print_obj_state(int tm_pct);
 
 #endif //BYLINS_SRC_GAME_MECHANICS_SIGHT_H_
